\name{inbio2.0}
\alias{inbio2.0}
\docType{package}
\title{
inbio (INcertidumbre BIOlogicos)
}
\description{
Conjunto de funciones para el cálculo de parámetros biológicos y su incertidumbre asociada mediante bootstrap no paramétrico.
}

\details{
\tabular{ll}{
Package: \tab inbio\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2013-02-18\cr
License: \tab GPL-2 \cr
}
}
\author{
Paz Sampedro
<paz.sampedro@co.ieo.es>

}
\references{
  Davison, A.C. & D.V. Hinkley. 1999. Bootstrap methods and their application. Cambridge University Press, 575 pp.

  Efron, B. 1979. Bootstrap methods: another look at the jackknife. Annals of Statistics, 7. 1-26 p.

  Gros, P., Laurec, A. & H. Dupouy. 1987. Estimation des parameters de croissance apport des techniques de reechantillonnage (Bootstrap & Jackniffe) et des statistiques robustes (Médianes et modes).  ICES Comité des statistiques,  CM 1987/D: 10, 25pp.

  Meyer, J.S., Ingersoll, C.G., McDonald, L.L. & M.S. Boyce. 1986. Estimating uncertainty in population growth rates: Jackknife vs Bootstrap techniques. Ecology, 67(5): 1156-1166.

  Moguedet, P. & N. Pérez. Estimation of Megrim (Lepidorhombus whiffiagonis) growth parameters, for males and females, from the ICES division VII: fitting to the V. Bertalanffy model using, resampling techniques, as well as several adjustable central values (mean, median and mode length at age). ICES CM 1988/G: 9. 18 pp.

  Quinn,T.J. & R.B. Deriso. 1999. Quantitative Fish Dynamics. Oxford University Press, 542 pp.

  Roa, R., Ernst, B & F. Tapia. 1999. Estimation of size at sexual maturity: an evaluation of analytical and resampling procedures. Fishery Bulletin 97: 570-580.

}

\keyword{ inbio2.0}
